refactor(ai-chat): extract pending-event revision branch (plan P8)
The three pending-event-revision sub-cases (clarify, unsupported_event_type,
apply) now live behind runPendingEventRevision(...). Returns
{ status: 'handled' | 'passthrough', fullContent } so the orchestrator
keeps owning the early-return + audit-finalize hand-off. Removes ~195 LOC
from handler.ts (899 -> 716). Snapshots 5/5; routes/ai/chat-handler.test
42/42.
